add automatical wrapping routes with type='switch'
